requirements.txt
mysql-connector-python>=8.0.32, <8.1.0
python-dotenv>=1.0.0, < 1.1.0